Seperate MediaData logic into its own module
