+++ src/bc/media_packet.rs
@@ -1,8 +1,28 @@
+use crate::bc::model::*;
+use crate::bc_protocol::connection::BcSubscription;
+use err_derive::Error;
 use log::trace;
+use log::*;
 use std::cmp::min;
+use std::collections::VecDeque;
 use std::convert::TryInto;
+use std::time::Duration;
+
+const INVALID_MEDIA_PACKETS: &[MediaDataKind] = &[
+    MediaDataKind::Invalid,
+    MediaDataKind::Continue,
+    MediaDataKind::Unknown,
+];
 
 pub const CHUNK_SIZE: usize = 40000;
+
+type Result<T> = std::result::Result<T, Error>;
+
+#[derive(Debug, Error)]
+pub enum Error {
+    #[error(display = "Timeout")]
+    Timeout(#[error(source)] std::sync::mpsc::RecvTimeoutError),
+}
 
 #[derive(Debug, PartialEq, Eq, Hash, Copy, Clone)]
 pub enum MediaDataKind {
@@ -149,5 +169,136 @@ impl MediaData {
 
     pub fn as_slice(&self) -> &[u8] {
         self.data.as_slice()
+    }
+}
+
+pub struct MediaDataSubscriber<'a> {
+    binary_buffer: VecDeque<u8>,
+    bc_sub: &'a BcSubscription<'a>,
+}
+
+impl<'a> MediaDataSubscriber<'a> {
+    pub fn from_bc_sub<'b>(bc_sub: &'b BcSubscription) -> MediaDataSubscriber<'b> {
+        MediaDataSubscriber {
+            binary_buffer: VecDeque::new(),
+            bc_sub: bc_sub,
+        }
+    }
+
+    fn fill_binary_buffer(&mut self, rx_timeout: Duration) -> Result<()> {
+        // Loop messages until we get binary add that data and return
+        loop {
+            let msg = self.bc_sub.rx.recv_timeout(rx_timeout)?;
+            if let BcBody::ModernMsg(ModernMsg {
+                binary: Some(binary),
+                ..
+            }) = msg.body
+            {
+                // Add the new binary to the buffer and return
+                self.binary_buffer.extend(binary);
+                break;
+            }
+        }
+        Ok(())
+    }
+
+    fn advance_to_media_packet(&mut self, rx_timeout: Duration) -> Result<()> {
+        // In the event we get an unknown packet we advance by brute force
+        // reading of bytes to the next valid magic
+        trace!("Advancing to next know packet header");
+        while self.binary_buffer.len() < 4 {
+            self.fill_binary_buffer(rx_timeout)?;
+        }
+
+        // Check the kind, if its invalid use pop a byte and try again
+        let mut magic = MediaDataSubscriber::get_first_n_deque(&self.binary_buffer, 4);
+        while INVALID_MEDIA_PACKETS.contains(&MediaData::kind_from_raw(&magic)) {
+            self.binary_buffer.pop_front();
+            while self.binary_buffer.len() < 4 {
+                self.fill_binary_buffer(rx_timeout)?;
+            }
+            magic = MediaDataSubscriber::get_first_n_deque(&self.binary_buffer, 4);
+        }
+
+        Ok(())
+    }
+
+    fn get_first_n_deque<T: std::clone::Clone>(deque: &VecDeque<T>, n: usize) -> Vec<T> {
+        // NOTE: I want to use make_contiguous
+        // This will make this func unneeded as we can use
+        // make_contiguous then as_slices.0
+        // We won't need the clone in this case either.
+        // This is an experimental feature.
+        // It is about to be moved to stable though
+        // As can be seen from this PR
+        // https://github.com/rust-lang/rust/pull/74559
+        let slice0 = deque.as_slices().0;
+        let slice1 = deque.as_slices().1;
+        if slice0.len() >= n {
+            return slice0[0..n].iter().cloned().collect();
+        } else {
+            let remain = n - slice0.len();
+            return slice0.iter().chain(&slice1[0..remain]).cloned().collect();
+        }
+    }
+
+    fn next_media_packet(&mut self, rx_timeout: Duration) -> std::result::Result<MediaData, Error> {
+        // Find the first packet (does nothing if already at one)
+        self.advance_to_media_packet(rx_timeout)?;
+
+        // Get the magic bytes (guaranteed by advance_to_media_packet)
+        let magic = MediaDataSubscriber::get_first_n_deque(&self.binary_buffer, 4);
+        let kind = MediaData::kind_from_raw(&magic);
+
+        // Get enough for the full header
+        let header_size = MediaData::header_size_from_raw(&magic);
+        while self.binary_buffer.len() < header_size {
+            self.fill_binary_buffer(rx_timeout)?;
+        }
+
+        // Get enough for the full data + 8 byte buffer
+        let header = MediaDataSubscriber::get_first_n_deque(&self.binary_buffer, header_size);
+        let data_size = MediaData::data_size_from_raw(&header);
+        let pad_size = MediaData::pad_size_from_raw(&header);
+        let full_size = header_size + data_size + pad_size;
+        while self.binary_buffer.len() < full_size {
+            self.fill_binary_buffer(rx_timeout)?;
+        }
+
+        // Pop the full binary buffer
+        let binary = self.binary_buffer.drain(..full_size);
+
+        Ok(MediaData {
+            data: binary.collect(),
+        })
+    }
+
+    pub fn get_media_packet_of_kind(
+        &mut self,
+        interested_kinds: &[MediaDataKind],
+        rx_timeout: Duration,
+    ) -> std::result::Result<MediaData, Error> {
+        trace!("Finding binary message of interest...");
+
+        let result_media_packet: MediaData;
+
+        // Loop over the messages until we find one we want
+        loop {
+            let media_packet = self.next_media_packet(rx_timeout)?;
+            match media_packet.kind() {
+                n if interested_kinds.contains(&n) => {
+                    result_media_packet = media_packet;
+                    break;
+                }
+                _ => {
+                    trace!(
+                        "Ignoring uninteresting binary data kind: {:?}",
+                        media_packet.kind()
+                    );
+                }
+            };
+        }
+
+        return Ok(result_media_packet);
     }
 }
